Refactor Day1 class and test for better organization and readability
This commit involves substantial refactoring of the Day1 class and associated test class in the Advent of Code 2023 project. The primary aim is to improve code readability, maintainability, and structure. Changes include the removal of lengthy commented lines and restructuring of the old 'Day1' class into 'Day01.Day01' which now resides in its dedicated Day01 package. Corresponding adjustment to the test class is also made. All resource files are moved under 'Day01' folder for better organization and easy access. Certain logic, such as replacing non-digit elements, are simplified for efficiency.